llm/model_fusion_project/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
aiohttp>=3.9.1
asyncio>=3.4.3
PyYAML>=5.4.1
torch>=2.1.0
transformers>=4.35.2
numpy>=1.24.0
scikit-learn>=1.3.0
prometheus_client>=0.11.0
psutil>=5.8.0
redis>=3.5.3
aioredis>=2.0.0
python-dateutil>=2.8.2
tenacity>=8.0.1
inquirer>=3.1.3
python-dotenv>=1.0.0
colorlog>=6.7.0
typing-extensions>=4.8.0